Move benchmarks to their own directory
